@@ -5,7 +5,7 @@ version: 0.0.4
 homepage: https://github.com/theiskaa/flutter_button
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.17.0"
 
 dependencies:
