@@ -4,15 +4,15 @@ class MenuFAB extends StatefulWidget {
   final Widget firstItem;
   final Widget seccondItem;
   final Widget thirdItem;
-  final AnimatedIconData animatedIcon;
-  final Color activeColor;
-  final Color inactiveColor;
-  final Curve curve;
+  final AnimatedIconData? animatedIcon;
+  final Color? activeColor;
+  final Color? inactiveColor;
+  final Curve? curve;
 
   MenuFAB({
-    @required this.firstItem,
-    @required this.seccondItem,
-    @required this.thirdItem,
+    required this.firstItem,
+    required this.seccondItem,
+    required this.thirdItem,
     this.animatedIcon,
     this.activeColor,
     this.inactiveColor,
@@ -25,10 +25,10 @@ class MenuFAB extends StatefulWidget {
 
 class _MenuFABState extends State<MenuFAB> with SingleTickerProviderStateMixin {
   bool isOpened = false;
-  AnimationController _animationController;
-  Animation<Color> _buttonColor;
-  Animation<double> _animateIcon;
-  Animation<double> _translateButton;
+  late AnimationController _animationController;
+  late Animation<Color?> _buttonColor;
+  late Animation<double> _animateIcon;
+  late Animation<double> _translateButton;
 
   @override
   initState() {
@@ -38,9 +38,8 @@ class _MenuFABState extends State<MenuFAB> with SingleTickerProviderStateMixin {
     _animateIcon =
         Tween<double>(begin: 0.0, end: 1.0).animate(_animationController);
     _buttonColor = ColorTween(
-      begin:
-          (widget.inactiveColor != null) ? widget.inactiveColor : Colors.blue,
-      end: (widget.activeColor != null) ? widget.activeColor : Colors.red,
+      begin: widget.inactiveColor ?? Colors.blue,
+      end: widget.activeColor ?? Colors.red,
     ).animate(CurvedAnimation(
       parent: _animationController,
       curve: Interval(
@@ -57,7 +56,7 @@ class _MenuFABState extends State<MenuFAB> with SingleTickerProviderStateMixin {
       curve: Interval(
         0.0,
         1.0,
-        curve: (widget.curve !=null) ? widget.curve : Curves.ease,
+        curve: widget.curve ?? Curves.ease,
       ),
     ));
 
@@ -101,7 +100,7 @@ class _MenuFABState extends State<MenuFAB> with SingleTickerProviderStateMixin {
         onPressed: animate,
         tooltip: 'Menu',
         child: AnimatedIcon(
-          icon: (widget.animatedIcon != null) ? widget.animatedIcon : AnimatedIcons.menu_close,
+          icon: widget.animatedIcon ?? AnimatedIcons.menu_close,
           progress: _animateIcon,
         ),
       ),
