@@ -5,24 +5,24 @@ import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 
 class GoogleAuthButton extends StatefulWidget {
   final VoidCallback onTap;
-  final String title;
-  final Color backgroundColor;
-  final Color titleColor;
-  final Color iconColor;
-  final BorderRadiusGeometry borderRadius;
-  final List<BoxShadow> shadows;
-  final double fontSize;
-  final double iconSize;
-  final FontWeight fontWeight;
-  final bool wOpacity;
-  final double opacityValue;
-  final bool wGradientColors;
-  final List<Color> gradientColors;
-  final AlignmentGeometry beginGradient;
-  final AlignmentGeometry endGradient;
+  final String? title;
+  final Color? backgroundColor;
+  final Color? titleColor;
+  final Color? iconColor;
+  final BorderRadiusGeometry? borderRadius;
+  final List<BoxShadow>? shadows;
+  final double? fontSize;
+  final double? iconSize;
+  final FontWeight? fontWeight;
+  final bool? wOpacity;
+  final double? opacityValue;
+  final bool? wGradientColors;
+  final List<Color>? gradientColors;
+  final AlignmentGeometry? beginGradient;
+  final AlignmentGeometry? endGradient;
 
   GoogleAuthButton({
-    @required this.onTap,
+    required this.onTap,
     this.title,
     this.backgroundColor,
     this.titleColor,
@@ -72,11 +72,7 @@ class _GoogleAuthButtonState extends State<GoogleAuthButton> {
 
   Opacity buildButtonWOpacity() {
     return Opacity(
-      opacity: _isTapped
-          ? (widget.opacityValue == null)
-              ? .7
-              : widget.opacityValue
-          : 1,
+      opacity: _isTapped ? widget.opacityValue ?? .7 : 1,
       child: buildButton(),
     );
   }
@@ -100,7 +96,7 @@ class _GoogleAuthButtonState extends State<GoogleAuthButton> {
 
             ///
             Text(
-              (widget.title != null) ? widget.title : "Google",
+              widget.title ?? "Google",
               style: TextStyle(
                 color: (widget.titleColor == null)
                     ? Color(0xff4285F4)
@@ -127,13 +123,9 @@ class _GoogleAuthButtonState extends State<GoogleAuthButton> {
       gradient: (widget.wGradientColors == true)
           ? (widget.gradientColors != null)
               ? LinearGradient(
-                  colors: widget.gradientColors,
-                  begin: (widget.beginGradient != null)
-                      ? widget.beginGradient
-                      : Alignment.topRight,
-                  end: (widget.endGradient != null)
-                      ? widget.beginGradient
-                      : Alignment.bottomLeft,
+                  colors: widget.gradientColors!,
+                  begin: widget.beginGradient ?? Alignment.topRight,
+                  end: widget.beginGradient ?? Alignment.bottomLeft,
                 )
               : LinearGradient(
                   begin: Alignment.topRight,
@@ -160,19 +152,19 @@ class _GoogleAuthButtonState extends State<GoogleAuthButton> {
 
 class CircularGGAuthButton extends StatefulWidget {
   final VoidCallback onTap;
-  final Color backgorundColor;
-  final Color iconColor;
-  final bool wOpacity;
-  final bool wBorder;
-  final Color borderColor;
-  final double opacityValue;
-  final double size;
-  final double iconSize;
-  final List<BoxShadow> shadows;
-  final BorderRadiusGeometry borderRadius;
+  final Color? backgorundColor;
+  final Color? iconColor;
+  final bool? wOpacity;
+  final bool? wBorder;
+  final Color? borderColor;
+  final double? opacityValue;
+  final double? size;
+  final double? iconSize;
+  final List<BoxShadow>? shadows;
+  final BorderRadiusGeometry? borderRadius;
 
   CircularGGAuthButton({
-    @required this.onTap,
+    required this.onTap,
     this.backgorundColor,
     this.iconColor,
     this.wOpacity,
@@ -214,11 +206,7 @@ class CircularGGAuthButtonState extends State<CircularGGAuthButton> {
       child: (widget.wOpacity == null || widget.wOpacity == false)
           ? body()
           : Opacity(
-              opacity: _isTapped
-                  ? (widget.opacityValue != null)
-                      ? widget.opacityValue
-                      : .7
-                  : 1,
+              opacity: _isTapped ? widget.opacityValue ?? .7 : 1,
               child: body(),
             ),
     );
@@ -252,9 +240,7 @@ class CircularGGAuthButtonState extends State<CircularGGAuthButton> {
           : Colors.white,
       border: (widget.wBorder != null && widget.wBorder != false)
           ? Border.all(
-              color: (widget.borderColor == null)
-                  ? Colors.black
-                  : widget.borderColor,
+              color: widget.borderColor ?? Colors.black,
             )
           : null,
       boxShadow: (widget.shadows == null) ? [] : widget.shadows,
@@ -270,13 +256,13 @@ class CircularGGAuthButtonState extends State<CircularGGAuthButton> {
 
 class MagicalGGButton extends StatefulWidget {
   final String title;
-  final double opacityValue;
+  final double? opacityValue;
   final VoidCallback onTap;
 
   MagicalGGButton({
-    @required this.title,
+    required this.title,
     this.opacityValue,
-    @required this.onTap,
+    required this.onTap,
   });
 
   @override
@@ -305,11 +291,7 @@ class _MagicalGGButtonState extends State<MagicalGGButton> {
         });
       },
       child: Opacity(
-        opacity: _isTapped
-            ? (widget.opacityValue != null)
-                ? widget.opacityValue
-                : .7
-            : 1,
+        opacity: _isTapped ? widget.opacityValue ?? .7 : 1,
         child: buildContainer(context),
       ),
     );
